% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_test_data}
\alias{generate_test_data}
\title{Generate Test Data}
\usage{
generate_test_data(N = 2000)
}
\arguments{
\item{N}{the number of the records}
}
\value{
test data that includes binary, numeric and multi-level factor variable
}
\description{
Generate Test Data
}
