% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortest.R
\name{ortest_function}
\alias{ortest_function}
\alias{ortest}
\title{Super-Learning Based Conditional Independence Test}
\usage{
ortest(x, y, S, suffStat)
}
\arguments{
\item{x}{The position of the exposure variable}

\item{y}{The position of the outcome variable}

\item{S}{The position of the conditional variable set}

\item{suffStat}{A list with elements needed for the test:
\describe{
\item{"dat"}{Data matrix where columns represent variables.}
\item{"method"}{The method used for the test: "sl", "linear", "linear_int"}
\item{"sl"}{Specifies the super learning model. If \code{y} is numeric, the default method is linear regression model, mean response,MARS, random forest and XGBoost. If \code{y} is binary, the default method is LDA, mean response,MARS, random forest and XGBoost.}
\item{"cross_fitting"}{Logical; whether cross-fitting is used in super learning.}
\item{"error_catch"}{Logical; whether using glm to refit the model when some folds return abnormal estimation}
\item{"kfolds"}{Number of folds for cross-fitting.}
\item{"two_way"}{Logical; whether to include two-way interactions between variables in \code{S}.}
\item{"three_way"}{Logical; whether to include three-way interactions between variables in \code{S}.}
}}
}
\value{
p-value for model \code{y ~ x + S}.
}
\description{
We test whether \code{x} and \code{y} are conditionally associated, given
\code{S}, using specific methods.
}
\details{
This function test whether \code{x} and \code{y} is independent conditional on
\code{S}. The final result only includes p value for model \code{y ~ x + S}.
}
