% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_hat.R
\name{psi.hat_linear}
\alias{psi.hat_linear}
\title{Regression Based Conditional Independence Test}
\usage{
psi.hat_linear(y, x, S=c(), subset = NULL, out.bin = TRUE, exp.bin = FALSE, root = c("uni","multi"))
}
\arguments{
\item{y}{Outcome variable, either binary or numeric}

\item{x}{Exposure variable, either binary or numeric}

\item{S}{Conditional variable set, can be empty}

\item{subset}{Optionally, control the subset of the dataset to be used}

\item{out.bin}{A logical evaluating to TRUE or FALSE indicating whether outcome variable is binary.}

\item{exp.bin}{A logical evaluating to TRUE or FALSE indicating whether exposure variable is binary.}

\item{root}{String specifying the method of solving equations}
}
\value{
Test statistics and standard error of the test.
}
\description{
We test whether \code{x} and \code{y} are conditionally associated, given
\code{S} using a generalized linear model.
}
\details{
All included variables should be either numeric or binary. If
\code{S} includes less than 4 variables, regression model is recommended. If
\code{y} is numeric, a linear regression model is fitted. If \code{y} is
binary, a logistical regression model is fitted. \code{x} and \code{S} are
included as explanatory variables.
This model is tested whether \code{x} and \code{y} is independent conditional on
\code{S}. The final result is the test statistics and standard error.
}
