% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_hat.R
\name{psi_hat_linear_int}
\alias{psi_hat_linear_int}
\title{Regression Based With Interaction Conditional Independence Test}
\usage{
psi_hat_linear_int(y, x, S=c(), subset = NULL, out_bin = TRUE, exp_bin = FALSE,
 two_way=FALSE, three_way=FALSE)
}
\arguments{
\item{y}{Outcome variable, either binary or numeric}

\item{x}{Exposure variable, either binary or numeric}

\item{S}{Conditional variable set, can be empty}

\item{subset}{Optionally, control the subset of the dataset to be used}

\item{out_bin}{A logical evaluating to TRUE or FALSE indicating whether outcome variable is binary.}

\item{exp_bin}{A logical evaluating to TRUE or FALSE indicating whether exposure variable is binary.}

\item{two_way}{A logical evaluating to TRUE or FALSE indicating whether two-way interaction for S is included in regression model.}

\item{three_way}{A logical evaluating to TRUE or FALSE indicating whether three-way interaction for S is included in regression model.}
}
\value{
Test statistics and standard error of the test.
}
\description{
We test whether \code{x} and \code{y} are conditionally associated, given
\code{S} using a generalized linear model allowing interaction between variables in \code{S}.
}
\details{
All included variables should be either numeric or binary. If
\code{S} includes less than 4 variables, regression model is recommended. If
\code{y} is numeric, a linear regression model is fitted. If \code{y} is
binary, a logistical regression model is fitted. \code{x} and \code{S} are
included as explanatory variables.
This model is tested whether \code{x} and \code{y} is independent conditional on
\code{S}. The final result is the test statistics and standard error.
}
