% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_hat.R
\name{psi_hat_sl}
\alias{psi_hat_sl}
\title{Super-Learning Based Conditional Independence Test}
\usage{
psi_hat_sl(
  y,
  x,
  S = c(),
  subset = NULL,
  out_bin = TRUE,
  exp_bin = FALSE,
  sl = NULL,
  cross_fitting = FALSE,
  kfolds = 5
)
}
\arguments{
\item{y}{Outcome variable, either binary or numeric}

\item{x}{Exposure variable, either binary or numeric}

\item{S}{Conditional variable set, can be empty}

\item{subset}{Optionally, control the subset of the dataset to be used}

\item{out_bin}{A logical evaluating to TRUE or FALSE indicating whether outcome variable is binary.}

\item{exp_bin}{A logical evaluating to TRUE or FALSE indicating whether exposure variable is binary.}

\item{sl}{Character string specifying models applied in super learning}

\item{cross_fitting}{A logical evaluating to TRUE or FALSE indicating whether cross-fitting is used.}

\item{kfolds}{A numeric indicating how many folds is used for cross-fitting if applied}
}
\value{
Test statistics and standard error of the test.
}
\description{
We test whether \code{x} and \code{y} are conditionally associated, given
\code{S} using a generalized linear model.
}
\details{
All included variables should be either numeric or binary. If \code{y} is numeric, the default method
is linear regression model, MARS, random forest and XGBoost. If
\code{y} is binary, the default method is linear regression model, MARS, random forest
and XGBoost.
This model is tested whether \code{x} and \code{y} is independent conditional on
\code{S}. The final result is the test statistics and standard error.
}
