% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortest_package.R
\docType{package}
\name{ortest-package}
\alias{ortest-package}
\title{Super-Learning Based Conditional Independence Test Package}
\description{
Provides conditional independence tests for mixed data types based on estimating conditional odds-ratio parameters. Implementations include regression-based estimating equations and optional Super Learner-based nuisance estimation with (optional) cross-fitting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chaoqiw0324/ortest}
  \item Report bugs at \url{https://github.com/chaoqiw0324/ortest/issues}
}

}
\author{
\strong{Maintainer}: Chaoqi Wu \email{cw3370@cumc.columbia.edu} [copyright holder]

}
\keyword{internal}
