% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortest-package.R
\docType{package}
\name{ortest-package}
\alias{ortest}
\alias{ortest-package}
\title{ortest: The Title of the Project}
\description{
Function for conditional independence test
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chaoqiw0324/ortest}
  \item Report bugs at \url{https://github.com/chaoqiw0324/ortest/issues}
}

}
\author{
\strong{Maintainer}: Chaoqi Wu \email{cw3370@cumc.columbia.edu} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
