% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORtest.R
\name{ORtest}
\alias{ORtest}
\title{Conditional Independence Test
We test whether \code{x} and \code{y} are conditionally associated, given
\code{S}}
\usage{
ORtest(x, y, S, suffStat)
}
\arguments{
\item{x}{The position of the exposure variable}

\item{y}{The position of the exposure variable}

\item{S}{The position of the conditional variable set}

\item{suffStat}{A list with four elements, "dat" is the dat matrix ,"sl" specifies the super learning model,"cross_fitting" indicates whether super learning is used and "kfolds" decides the folds for cross-fitting.}
}
\value{
p-value for model \code{y ~ x + S} and \code{x ~ y+S}.
}
\description{
Conditional Independence Test
We test whether \code{x} and \code{y} are conditionally associated, given
\code{S}
}
\details{
This function test whether \code{x} and \code{y} is independent conditional on
\code{S}. The final results includes p-value for model \code{y ~ x + S}
and \code{x ~ y+S}.
}
