% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_hat.R
\name{psi_hat_linear}
\alias{psi_hat_linear}
\title{Regression Based Conditional Independence Test}
\usage{
psi_hat_linear(y, x, S = c(), subset = NULL, out_bin = TRUE, exp_bin = FALSE)
}
\arguments{
\item{y}{Outcome variable, either binary or numeric}

\item{x}{Exposure variable, either binary or numeric}

\item{S}{Conditional variable set, can be empty}

\item{subset}{Optionally, control the subset of the dataset to be used}

\item{out_bin}{A logical evaluating to TRUE or FALSE indicating whether
outcome variable is binary.}

\item{exp_bin}{A logical evaluating to TRUE or FALSE indicating whether
exposure variable is binary.}
}
\value{
Test statistics and standard error of the test.
}
\description{
We test whether \code{x} and \code{y} are conditionally associated, given
\code{S} using a generalized linear model.
}
\details{
This model is tested whether \code{x} and \code{y} is independent conditional on
\code{S}. The final result is the test statistics and standard error.
}
